+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Crystal Nest Semantic Versioning](https://crystalne
 
 ## [Unreleased]
 
-- Nothing new.
+- Removed spectator setting in favor of the new `RenderMode`.
+- Add new `RenderMode` to choose between normal mode, invisible mode, spectator mode, and ghost mode.
 
 ## [v0.2.0] - 2025/04/11